[Refactor] Move query timeout hook out of runtime
